WIP: restructuration, testing
