+++ tradetrust/ts-document-store-worker/src/constants.ts
@@ -1,21 +1,15 @@
 const OPEN_ATTESTATION_VERSION_ID_V2_SHORT: string = 'open-attestation/2.0';
 const OPEN_ATTESTATION_VERSION_ID_V3_SHORT: string = 'open-attestation/3.0';
-const OPEN_ATTESTATION_VERSION_ID_V2_FRAMEWORK: string = 'https://schema.openattestation.com/2.0/schema.json';
-const OPEN_ATTESTATION_VERSION_ID_V3_FRAMEWORK: string = 'https://schema.openattestation.com/3.0/schema.json';
 const DOCUMENT_STORE_PROOF_TYPE: string = 'DOCUMENT_STORE';
 
 export {
   OPEN_ATTESTATION_VERSION_ID_V2_SHORT,
   OPEN_ATTESTATION_VERSION_ID_V3_SHORT,
-  OPEN_ATTESTATION_VERSION_ID_V2_FRAMEWORK,
-  OPEN_ATTESTATION_VERSION_ID_V3_FRAMEWORK,
   DOCUMENT_STORE_PROOF_TYPE
 }
 
 export default {
   DOCUMENT_STORE_PROOF_TYPE,
   OPEN_ATTESTATION_VERSION_ID_V2_SHORT,
-  OPEN_ATTESTATION_VERSION_ID_V3_SHORT,
-  OPEN_ATTESTATION_VERSION_ID_V2_FRAMEWORK,
-  OPEN_ATTESTATION_VERSION_ID_V3_FRAMEWORK
+  OPEN_ATTESTATION_VERSION_ID_V3_SHORT
 }
