+++ tradetrust/ts-document-store-worker/src/logger.ts
@@ -1,0 +1,5 @@
+import pino from 'pino';
+
+const logger = pino({level: 'debug'});
+
+export { logger };
