+++ tradetrust/ts-document-store-worker/src/tasks/index.ts
@@ -1,0 +1,16 @@
+import ComposeBatch from './compose-batch';
+import WrapBatch from './wrap-batch';
+import IssueBatch from './issue-batch';
+import SaveIssuedBatch from './save-issued-batch';
+import ProcessDocuments from './process-documents';
+import { Batch } from './data';
+
+
+export {
+  Batch,
+  ComposeBatch,
+  WrapBatch,
+  IssueBatch,
+  SaveIssuedBatch,
+  ProcessDocuments
+}
