+++ tradetrust/ts-document-store-worker/src/tasks/interfaces.ts
@@ -1,0 +1,16 @@
+interface UnwrappedDocument{
+  size: number,
+  body: any
+}
+
+
+interface Task<T>{
+  start(): T;
+  next(): void
+}
+
+
+export {
+  Task,
+  UnwrappedDocument
+}
