+++ tradetrust/ts-document-store-worker/tests/process-documents-task.test.ts
@@ -1,0 +1,3 @@
+describe('Test', ()=>{
+  test('test', ()=>{});
+})
